rfp-assistant/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
pydantic==1.10.7
boto3==1.26.133
langchain==0.0.267
langchain-openai==0.0.5
langchain-community==0.0.5
langchain-text-splitters==0.0.1
openai==1.2.0
PyPDF2==3.0.1
python-docx==0.8.11
faiss-cpu==1.7.4
python-dotenv==1.0.0
jinja2==3.1.2
numpy==1.24.3
